+++ lib/forge-std/lib/ds-test/src/test.sol
@@ -16,42 +16,46 @@
 pragma solidity >=0.4.23;
 
 contract DSTest {
-    event log                    (string);
-    event logs                   (bytes);
+    event log(string);
+    event logs(bytes);
 
-    event log_address            (address);
-    event log_bytes32            (bytes32);
-    event log_int                (int);
-    event log_uint               (uint);
-    event log_bytes              (bytes);
-    event log_string             (string);
+    event log_address(address);
+    event log_bytes32(bytes32);
+    event log_int(int256);
+    event log_uint(uint256);
+    event log_bytes(bytes);
+    event log_string(string);
 
-    event log_named_address      (string key, address val);
-    event log_named_bytes32      (string key, bytes32 val);
-    event log_named_decimal_int  (string key, int val, uint decimals);
-    event log_named_decimal_uint (string key, uint val, uint decimals);
-    event log_named_int          (string key, int val);
-    event log_named_uint         (string key, uint val);
-    event log_named_bytes        (string key, bytes val);
-    event log_named_string       (string key, string val);
+    event log_named_address(string key, address val);
+    event log_named_bytes32(string key, bytes32 val);
+    event log_named_decimal_int(string key, int256 val, uint256 decimals);
+    event log_named_decimal_uint(string key, uint256 val, uint256 decimals);
+    event log_named_int(string key, int256 val);
+    event log_named_uint(string key, uint256 val);
+    event log_named_bytes(string key, bytes val);
+    event log_named_string(string key, string val);
 
     bool public IS_TEST = true;
     bool public failed;
 
     address constant HEVM_ADDRESS =
-        address(bytes20(uint160(uint256(keccak256('hevm cheat code')))));
+        address(bytes20(uint160(uint256(keccak256("hevm cheat code")))));
 
-    modifier mayRevert() { _; }
-    modifier testopts(string memory) { _; }
+    modifier mayRevert() {
+        _;
+    }
+    modifier testopts(string memory) {
+        _;
+    }
 
     function fail() internal {
         failed = true;
     }
 
     modifier logs_gas() {
-        uint startGas = gasleft();
+        uint256 startGas = gasleft();
         _;
-        uint endGas = gasleft();
+        uint256 endGas = gasleft();
         emit log_named_uint("gas", startGas - endGas);
     }
 
@@ -77,9 +81,14 @@ contract DSTest {
             fail();
         }
     }
-    function assertEq(address a, address b, string memory err) internal {
+
+    function assertEq(
+        address a,
+        address b,
+        string memory err
+    ) internal {
         if (a != b) {
-            emit log_named_string ("Error", err);
+            emit log_named_string("Error", err);
             assertEq(a, b);
         }
     }
@@ -92,20 +101,31 @@ contract DSTest {
             fail();
         }
     }
-    function assertEq(bytes32 a, bytes32 b, string memory err) internal {
+
+    function assertEq(
+        bytes32 a,
+        bytes32 b,
+        string memory err
+    ) internal {
         if (a != b) {
-            emit log_named_string ("Error", err);
+            emit log_named_string("Error", err);
             assertEq(a, b);
         }
     }
+
     function assertEq32(bytes32 a, bytes32 b) internal {
         assertEq(a, b);
     }
-    function assertEq32(bytes32 a, bytes32 b, string memory err) internal {
+
+    function assertEq32(
+        bytes32 a,
+        bytes32 b,
+        string memory err
+    ) internal {
         assertEq(a, b, err);
     }
 
-    function assertEq(int a, int b) internal {
+    function assertEq(int256 a, int256 b) internal {
         if (a != b) {
             emit log("Error: a == b not satisfied [int]");
             emit log_named_int("  Expected", b);
@@ -113,13 +133,19 @@ contract DSTest {
             fail();
         }
     }
-    function assertEq(int a, int b, string memory err) internal {
+
+    function assertEq(
+        int256 a,
+        int256 b,
+        string memory err
+    ) internal {
         if (a != b) {
             emit log_named_string("Error", err);
             assertEq(a, b);
         }
     }
-    function assertEq(uint a, uint b) internal {
+
+    function assertEq(uint256 a, uint256 b) internal {
         if (a != b) {
             emit log("Error: a == b not satisfied [uint]");
             emit log_named_uint("  Expected", b);
@@ -127,13 +153,23 @@ contract DSTest {
             fail();
         }
     }
-    function assertEq(uint a, uint b, string memory err) internal {
+
+    function assertEq(
+        uint256 a,
+        uint256 b,
+        string memory err
+    ) internal {
         if (a != b) {
             emit log_named_string("Error", err);
             assertEq(a, b);
         }
     }
-    function assertEqDecimal(int a, int b, uint decimals) internal {
+
+    function assertEqDecimal(
+        int256 a,
+        int256 b,
+        uint256 decimals
+    ) internal {
         if (a != b) {
             emit log("Error: a == b not satisfied [decimal int]");
             emit log_named_decimal_int("  Expected", b, decimals);
@@ -141,13 +177,24 @@ contract DSTest {
             fail();
         }
     }
-    function assertEqDecimal(int a, int b, uint decimals, string memory err) internal {
+
+    function assertEqDecimal(
+        int256 a,
+        int256 b,
+        uint256 decimals,
+        string memory err
+    ) internal {
         if (a != b) {
             emit log_named_string("Error", err);
             assertEqDecimal(a, b, decimals);
         }
     }
-    function assertEqDecimal(uint a, uint b, uint decimals) internal {
+
+    function assertEqDecimal(
+        uint256 a,
+        uint256 b,
+        uint256 decimals
+    ) internal {
         if (a != b) {
             emit log("Error: a == b not satisfied [decimal uint]");
             emit log_named_decimal_uint("  Expected", b, decimals);
@@ -155,14 +202,20 @@ contract DSTest {
             fail();
         }
     }
-    function assertEqDecimal(uint a, uint b, uint decimals, string memory err) internal {
+
+    function assertEqDecimal(
+        uint256 a,
+        uint256 b,
+        uint256 decimals,
+        string memory err
+    ) internal {
         if (a != b) {
             emit log_named_string("Error", err);
             assertEqDecimal(a, b, decimals);
         }
     }
 
-    function assertGt(uint a, uint b) internal {
+    function assertGt(uint256 a, uint256 b) internal {
         if (a <= b) {
             emit log("Error: a > b not satisfied [uint]");
             emit log_named_uint("  Value a", a);
@@ -170,13 +223,19 @@ contract DSTest {
             fail();
         }
     }
-    function assertGt(uint a, uint b, string memory err) internal {
+
+    function assertGt(
+        uint256 a,
+        uint256 b,
+        string memory err
+    ) internal {
         if (a <= b) {
             emit log_named_string("Error", err);
             assertGt(a, b);
         }
     }
-    function assertGt(int a, int b) internal {
+
+    function assertGt(int256 a, int256 b) internal {
         if (a <= b) {
             emit log("Error: a > b not satisfied [int]");
             emit log_named_int("  Value a", a);
@@ -184,13 +243,23 @@ contract DSTest {
             fail();
         }
     }
-    function assertGt(int a, int b, string memory err) internal {
+
+    function assertGt(
+        int256 a,
+        int256 b,
+        string memory err
+    ) internal {
         if (a <= b) {
             emit log_named_string("Error", err);
             assertGt(a, b);
         }
     }
-    function assertGtDecimal(int a, int b, uint decimals) internal {
+
+    function assertGtDecimal(
+        int256 a,
+        int256 b,
+        uint256 decimals
+    ) internal {
         if (a <= b) {
             emit log("Error: a > b not satisfied [decimal int]");
             emit log_named_decimal_int("  Value a", a, decimals);
@@ -198,13 +267,24 @@ contract DSTest {
             fail();
         }
     }
-    function assertGtDecimal(int a, int b, uint decimals, string memory err) internal {
+
+    function assertGtDecimal(
+        int256 a,
+        int256 b,
+        uint256 decimals,
+        string memory err
+    ) internal {
         if (a <= b) {
             emit log_named_string("Error", err);
             assertGtDecimal(a, b, decimals);
         }
     }
-    function assertGtDecimal(uint a, uint b, uint decimals) internal {
+
+    function assertGtDecimal(
+        uint256 a,
+        uint256 b,
+        uint256 decimals
+    ) internal {
         if (a <= b) {
             emit log("Error: a > b not satisfied [decimal uint]");
             emit log_named_decimal_uint("  Value a", a, decimals);
@@ -212,14 +292,20 @@ contract DSTest {
             fail();
         }
     }
-    function assertGtDecimal(uint a, uint b, uint decimals, string memory err) internal {
+
+    function assertGtDecimal(
+        uint256 a,
+        uint256 b,
+        uint256 decimals,
+        string memory err
+    ) internal {
         if (a <= b) {
             emit log_named_string("Error", err);
             assertGtDecimal(a, b, decimals);
         }
     }
 
-    function assertGe(uint a, uint b) internal {
+    function assertGe(uint256 a, uint256 b) internal {
         if (a < b) {
             emit log("Error: a >= b not satisfied [uint]");
             emit log_named_uint("  Value a", a);
@@ -227,13 +313,19 @@ contract DSTest {
             fail();
         }
     }
-    function assertGe(uint a, uint b, string memory err) internal {
+
+    function assertGe(
+        uint256 a,
+        uint256 b,
+        string memory err
+    ) internal {
         if (a < b) {
             emit log_named_string("Error", err);
             assertGe(a, b);
         }
     }
-    function assertGe(int a, int b) internal {
+
+    function assertGe(int256 a, int256 b) internal {
         if (a < b) {
             emit log("Error: a >= b not satisfied [int]");
             emit log_named_int("  Value a", a);
@@ -241,13 +333,23 @@ contract DSTest {
             fail();
         }
     }
-    function assertGe(int a, int b, string memory err) internal {
+
+    function assertGe(
+        int256 a,
+        int256 b,
+        string memory err
+    ) internal {
         if (a < b) {
             emit log_named_string("Error", err);
             assertGe(a, b);
         }
     }
-    function assertGeDecimal(int a, int b, uint decimals) internal {
+
+    function assertGeDecimal(
+        int256 a,
+        int256 b,
+        uint256 decimals
+    ) internal {
         if (a < b) {
             emit log("Error: a >= b not satisfied [decimal int]");
             emit log_named_decimal_int("  Value a", a, decimals);
@@ -255,13 +357,24 @@ contract DSTest {
             fail();
         }
     }
-    function assertGeDecimal(int a, int b, uint decimals, string memory err) internal {
+
+    function assertGeDecimal(
+        int256 a,
+        int256 b,
+        uint256 decimals,
+        string memory err
+    ) internal {
         if (a < b) {
             emit log_named_string("Error", err);
             assertGeDecimal(a, b, decimals);
         }
     }
-    function assertGeDecimal(uint a, uint b, uint decimals) internal {
+
+    function assertGeDecimal(
+        uint256 a,
+        uint256 b,
+        uint256 decimals
+    ) internal {
         if (a < b) {
             emit log("Error: a >= b not satisfied [decimal uint]");
             emit log_named_decimal_uint("  Value a", a, decimals);
@@ -269,14 +382,20 @@ contract DSTest {
             fail();
         }
     }
-    function assertGeDecimal(uint a, uint b, uint decimals, string memory err) internal {
+
+    function assertGeDecimal(
+        uint256 a,
+        uint256 b,
+        uint256 decimals,
+        string memory err
+    ) internal {
         if (a < b) {
             emit log_named_string("Error", err);
             assertGeDecimal(a, b, decimals);
         }
     }
 
-    function assertLt(uint a, uint b) internal {
+    function assertLt(uint256 a, uint256 b) internal {
         if (a >= b) {
             emit log("Error: a < b not satisfied [uint]");
             emit log_named_uint("  Value a", a);
@@ -284,13 +403,19 @@ contract DSTest {
             fail();
         }
     }
-    function assertLt(uint a, uint b, string memory err) internal {
+
+    function assertLt(
+        uint256 a,
+        uint256 b,
+        string memory err
+    ) internal {
         if (a >= b) {
             emit log_named_string("Error", err);
             assertLt(a, b);
         }
     }
-    function assertLt(int a, int b) internal {
+
+    function assertLt(int256 a, int256 b) internal {
         if (a >= b) {
             emit log("Error: a < b not satisfied [int]");
             emit log_named_int("  Value a", a);
@@ -298,13 +423,23 @@ contract DSTest {
             fail();
         }
     }
-    function assertLt(int a, int b, string memory err) internal {
+
+    function assertLt(
+        int256 a,
+        int256 b,
+        string memory err
+    ) internal {
         if (a >= b) {
             emit log_named_string("Error", err);
             assertLt(a, b);
         }
     }
-    function assertLtDecimal(int a, int b, uint decimals) internal {
+
+    function assertLtDecimal(
+        int256 a,
+        int256 b,
+        uint256 decimals
+    ) internal {
         if (a >= b) {
             emit log("Error: a < b not satisfied [decimal int]");
             emit log_named_decimal_int("  Value a", a, decimals);
@@ -312,13 +447,24 @@ contract DSTest {
             fail();
         }
     }
-    function assertLtDecimal(int a, int b, uint decimals, string memory err) internal {
+
+    function assertLtDecimal(
+        int256 a,
+        int256 b,
+        uint256 decimals,
+        string memory err
+    ) internal {
         if (a >= b) {
             emit log_named_string("Error", err);
             assertLtDecimal(a, b, decimals);
         }
     }
-    function assertLtDecimal(uint a, uint b, uint decimals) internal {
+
+    function assertLtDecimal(
+        uint256 a,
+        uint256 b,
+        uint256 decimals
+    ) internal {
         if (a >= b) {
             emit log("Error: a < b not satisfied [decimal uint]");
             emit log_named_decimal_uint("  Value a", a, decimals);
@@ -326,14 +472,20 @@ contract DSTest {
             fail();
         }
     }
-    function assertLtDecimal(uint a, uint b, uint decimals, string memory err) internal {
+
+    function assertLtDecimal(
+        uint256 a,
+        uint256 b,
+        uint256 decimals,
+        string memory err
+    ) internal {
         if (a >= b) {
             emit log_named_string("Error", err);
             assertLtDecimal(a, b, decimals);
         }
     }
 
-    function assertLe(uint a, uint b) internal {
+    function assertLe(uint256 a, uint256 b) internal {
         if (a > b) {
             emit log("Error: a <= b not satisfied [uint]");
             emit log_named_uint("  Value a", a);
@@ -341,13 +493,19 @@ contract DSTest {
             fail();
         }
     }
-    function assertLe(uint a, uint b, string memory err) internal {
+
+    function assertLe(
+        uint256 a,
+        uint256 b,
+        string memory err
+    ) internal {
         if (a > b) {
             emit log_named_string("Error", err);
             assertLe(a, b);
         }
     }
-    function assertLe(int a, int b) internal {
+
+    function assertLe(int256 a, int256 b) internal {
         if (a > b) {
             emit log("Error: a <= b not satisfied [int]");
             emit log_named_int("  Value a", a);
@@ -355,13 +513,23 @@ contract DSTest {
             fail();
         }
     }
-    function assertLe(int a, int b, string memory err) internal {
+
+    function assertLe(
+        int256 a,
+        int256 b,
+        string memory err
+    ) internal {
         if (a > b) {
             emit log_named_string("Error", err);
             assertLe(a, b);
         }
     }
-    function assertLeDecimal(int a, int b, uint decimals) internal {
+
+    function assertLeDecimal(
+        int256 a,
+        int256 b,
+        uint256 decimals
+    ) internal {
         if (a > b) {
             emit log("Error: a <= b not satisfied [decimal int]");
             emit log_named_decimal_int("  Value a", a, decimals);
@@ -369,13 +537,24 @@ contract DSTest {
             fail();
         }
     }
-    function assertLeDecimal(int a, int b, uint decimals, string memory err) internal {
+
+    function assertLeDecimal(
+        int256 a,
+        int256 b,
+        uint256 decimals,
+        string memory err
+    ) internal {
         if (a > b) {
             emit log_named_string("Error", err);
             assertLeDecimal(a, b, decimals);
         }
     }
-    function assertLeDecimal(uint a, uint b, uint decimals) internal {
+
+    function assertLeDecimal(
+        uint256 a,
+        uint256 b,
+        uint256 decimals
+    ) internal {
         if (a > b) {
             emit log("Error: a <= b not satisfied [decimal uint]");
             emit log_named_decimal_uint("  Value a", a, decimals);
@@ -383,7 +562,13 @@ contract DSTest {
             fail();
         }
     }
-    function assertLeDecimal(uint a, uint b, uint decimals, string memory err) internal {
+
+    function assertLeDecimal(
+        uint256 a,
+        uint256 b,
+        uint256 decimals,
+        string memory err
+    ) internal {
         if (a > b) {
             emit log_named_string("Error", err);
             assertGeDecimal(a, b, decimals);
@@ -398,17 +583,26 @@ contract DSTest {
             fail();
         }
     }
-    function assertEq(string memory a, string memory b, string memory err) internal {
+
+    function assertEq(
+        string memory a,
+        string memory b,
+        string memory err
+    ) internal {
         if (keccak256(abi.encodePacked(a)) != keccak256(abi.encodePacked(b))) {
             emit log_named_string("Error", err);
             assertEq(a, b);
         }
     }
 
-    function checkEq0(bytes memory a, bytes memory b) internal pure returns (bool ok) {
+    function checkEq0(bytes memory a, bytes memory b)
+        internal
+        pure
+        returns (bool ok)
+    {
         ok = true;
         if (a.length == b.length) {
-            for (uint i = 0; i < a.length; i++) {
+            for (uint256 i = 0; i < a.length; i++) {
                 if (a[i] != b[i]) {
                     ok = false;
                 }
@@ -417,6 +611,7 @@ contract DSTest {
             ok = false;
         }
     }
+
     function assertEq0(bytes memory a, bytes memory b) internal {
         if (!checkEq0(a, b)) {
             emit log("Error: a == b not satisfied [bytes]");
@@ -425,7 +620,12 @@ contract DSTest {
             fail();
         }
     }
-    function assertEq0(bytes memory a, bytes memory b, string memory err) internal {
+
+    function assertEq0(
+        bytes memory a,
+        bytes memory b,
+        string memory err
+    ) internal {
         if (!checkEq0(a, b)) {
             emit log_named_string("Error", err);
             assertEq0(a, b);
