+++ lib/forge-std/src/test/StdStorage.t.sol
@@ -18,20 +18,30 @@ contract StdStorageTest is DSTest {
     }
 
     function testStorageHidden() public {
-        assertEq(uint256(keccak256("my.random.var")), stdstore.target(address(test)).sig("hidden()").find());
+        assertEq(
+            uint256(keccak256("my.random.var")),
+            stdstore.target(address(test)).sig("hidden()").find()
+        );
     }
 
     function testStorageObvious() public {
-        assertEq(uint256(0), stdstore.target(address(test)).sig("exists()").find());
+        assertEq(
+            uint256(0),
+            stdstore.target(address(test)).sig("exists()").find()
+        );
     }
 
     function testStorageCheckedWriteHidden() public {
-        stdstore.target(address(test)).sig(test.hidden.selector).checked_write(100);
+        stdstore.target(address(test)).sig(test.hidden.selector).checked_write(
+            100
+        );
         assertEq(uint256(test.hidden()), 100);
     }
 
     function testStorageCheckedWriteObvious() public {
-        stdstore.target(address(test)).sig(test.exists.selector).checked_write(100);
+        stdstore.target(address(test)).sig(test.exists.selector).checked_write(
+            100
+        );
         assertEq(test.exists(), 100);
     }
 
@@ -52,7 +62,7 @@ contract StdStorageTest is DSTest {
             .with_key(address(this))
             .depth(1)
             .find();
-       assertEq(uint256(keccak256(abi.encode(address(this), 4))) + 1, slot);
+        assertEq(uint256(keccak256(abi.encode(address(this), 4))) + 1, slot);
     }
 
     function testStorageDeepMap() public {
@@ -62,7 +72,17 @@ contract StdStorageTest is DSTest {
             .with_key(address(this))
             .with_key(address(this))
             .find();
-        assertEq(uint256(keccak256(abi.encode(address(this), keccak256(abi.encode(address(this), uint(5)))))), slot);
+        assertEq(
+            uint256(
+                keccak256(
+                    abi.encode(
+                        address(this),
+                        keccak256(abi.encode(address(this), uint256(5)))
+                    )
+                )
+            ),
+            slot
+        );
     }
 
     function testStorageCheckedWriteDeepMap() public {
@@ -83,7 +103,19 @@ contract StdStorageTest is DSTest {
             .with_key(address(this))
             .depth(0)
             .find();
-        assertEq(bytes32(uint256(keccak256(abi.encode(address(this), keccak256(abi.encode(address(this), uint(6)))))) + 0), bytes32(slot));
+        assertEq(
+            bytes32(
+                uint256(
+                    keccak256(
+                        abi.encode(
+                            address(this),
+                            keccak256(abi.encode(address(this), uint256(6)))
+                        )
+                    )
+                ) + 0
+            ),
+            bytes32(slot)
+        );
     }
 
     function testStorageDeepMapStructB() public {
@@ -94,7 +126,19 @@ contract StdStorageTest is DSTest {
             .with_key(address(this))
             .depth(1)
             .find();
-        assertEq(bytes32(uint256(keccak256(abi.encode(address(this), keccak256(abi.encode(address(this), uint(6)))))) + 1), bytes32(slot));
+        assertEq(
+            bytes32(
+                uint256(
+                    keccak256(
+                        abi.encode(
+                            address(this),
+                            keccak256(abi.encode(address(this), uint256(6)))
+                        )
+                    )
+                ) + 1
+            ),
+            bytes32(slot)
+        );
     }
 
     function testStorageCheckedWriteDeepMapStructA() public {
@@ -105,7 +149,10 @@ contract StdStorageTest is DSTest {
             .with_key(address(this))
             .depth(0)
             .checked_write(100);
-        (uint256 a, uint256 b) = test.deep_map_struct(address(this), address(this));
+        (uint256 a, uint256 b) = test.deep_map_struct(
+            address(this),
+            address(this)
+        );
         assertEq(100, a);
         assertEq(0, b);
     }
@@ -118,7 +165,10 @@ contract StdStorageTest is DSTest {
             .with_key(address(this))
             .depth(1)
             .checked_write(100);
-        (uint256 a, uint256 b) = test.deep_map_struct(address(this), address(this));
+        (uint256 a, uint256 b) = test.deep_map_struct(
+            address(this),
+            address(this)
+        );
         assertEq(0, a);
         assertEq(100, b);
     }
@@ -148,64 +198,111 @@ contract StdStorageTest is DSTest {
     }
 
     function testStorageStructA() public {
-        uint256 slot = stdstore.target(address(test)).sig(test.basic.selector).depth(0).find();
+        uint256 slot = stdstore
+            .target(address(test))
+            .sig(test.basic.selector)
+            .depth(0)
+            .find();
         assertEq(uint256(7), slot);
     }
 
     function testStorageStructB() public {
-        uint256 slot = stdstore.target(address(test)).sig(test.basic.selector).depth(1).find();
+        uint256 slot = stdstore
+            .target(address(test))
+            .sig(test.basic.selector)
+            .depth(1)
+            .find();
         assertEq(uint256(7) + 1, slot);
     }
 
     function testStorageCheckedWriteStructA() public {
-        stdstore.target(address(test)).sig(test.basic.selector).depth(0).checked_write(100);
+        stdstore
+            .target(address(test))
+            .sig(test.basic.selector)
+            .depth(0)
+            .checked_write(100);
         (uint256 a, uint256 b) = test.basic();
         assertEq(a, 100);
         assertEq(b, 1337);
     }
 
     function testStorageCheckedWriteStructB() public {
-         stdstore.target(address(test)).sig(test.basic.selector).depth(1).checked_write(100);
+        stdstore
+            .target(address(test))
+            .sig(test.basic.selector)
+            .depth(1)
+            .checked_write(100);
         (uint256 a, uint256 b) = test.basic();
         assertEq(a, 1337);
         assertEq(b, 100);
     }
 
     function testStorageMapAddrFound() public {
-        uint256 slot = stdstore.target(address(test)).sig(test.map_addr.selector).with_key(address(this)).find();
-        assertEq(uint256(keccak256(abi.encode(address(this), uint(1)))), slot);
+        uint256 slot = stdstore
+            .target(address(test))
+            .sig(test.map_addr.selector)
+            .with_key(address(this))
+            .find();
+        assertEq(
+            uint256(keccak256(abi.encode(address(this), uint256(1)))),
+            slot
+        );
     }
 
     function testStorageMapUintFound() public {
-        uint256 slot = stdstore.target(address(test)).sig(test.map_uint.selector).with_key(100).find();
-        assertEq(uint256(keccak256(abi.encode(100, uint(2)))), slot);
+        uint256 slot = stdstore
+            .target(address(test))
+            .sig(test.map_uint.selector)
+            .with_key(100)
+            .find();
+        assertEq(uint256(keccak256(abi.encode(100, uint256(2)))), slot);
     }
 
     function testStorageCheckedWriteMapUint() public {
-        stdstore.target(address(test)).sig(test.map_uint.selector).with_key(100).checked_write(100);
+        stdstore
+            .target(address(test))
+            .sig(test.map_uint.selector)
+            .with_key(100)
+            .checked_write(100);
         assertEq(100, test.map_uint(100));
     }
 
     function testStorageCheckedWriteMapAddr() public {
-        stdstore.target(address(test)).sig(test.map_addr.selector).with_key(address(this)).checked_write(100);
+        stdstore
+            .target(address(test))
+            .sig(test.map_addr.selector)
+            .with_key(address(this))
+            .checked_write(100);
         assertEq(100, test.map_addr(address(this)));
     }
 
     function testStorageCheckedWriteMapBool() public {
-        stdstore.target(address(test)).sig(test.map_bool.selector).with_key(address(this)).checked_write(true);
+        stdstore
+            .target(address(test))
+            .sig(test.map_bool.selector)
+            .with_key(address(this))
+            .checked_write(true);
         assertTrue(test.map_bool(address(this)));
     }
 
     function testFailStorageCheckedWriteMapPacked() public {
         // expect PackedSlot error but not external call so cant expectRevert
-        stdstore.target(address(test)).sig(test.read_struct_lower.selector).with_key(address(uint160(1337))).checked_write(100);
+        stdstore
+            .target(address(test))
+            .sig(test.read_struct_lower.selector)
+            .with_key(address(uint160(1337)))
+            .checked_write(100);
     }
 
     function testStorageCheckedWriteMapPackedSuccess() public {
         uint256 full = test.map_packed(address(1337));
         // keep upper 128, set lower 128 to 1337
         full = (full & (uint256((1 << 128) - 1) << 128)) | 1337;
-        stdstore.target(address(test)).sig(test.map_packed.selector).with_key(address(uint160(1337))).checked_write(full);
+        stdstore
+            .target(address(test))
+            .sig(test.map_packed.selector)
+            .with_key(address(uint160(1337)))
+            .checked_write(full);
         assertEq(1337, test.read_struct_lower(address(1337)));
     }
 
@@ -217,7 +314,7 @@ contract StdStorageTest is DSTest {
     function testFailStorageNativePack() public {
         stdstore.target(address(test)).sig(test.tA.selector).find();
         stdstore.target(address(test)).sig(test.tB.selector).find();
-        
+
         // these both would fail
         stdstore.target(address(test)).sig(test.tC.selector).find();
         stdstore.target(address(test)).sig(test.tD.selector).find();
@@ -231,16 +328,15 @@ contract StorageTest {
     mapping(address => uint256) public map_packed;
     mapping(address => UnpackedStruct) public map_struct;
     mapping(address => mapping(address => uint256)) public deep_map;
-    mapping(address => mapping(address => UnpackedStruct)) public deep_map_struct;
+    mapping(address => mapping(address => UnpackedStruct))
+        public deep_map_struct;
     UnpackedStruct public basic;
 
     uint248 public tA;
     bool public tB;
 
-
     bool public tC = false;
-    uint248 public tD = 1;    
-
+    uint248 public tD = 1;
 
     struct UnpackedStruct {
         uint256 a;
@@ -250,14 +346,11 @@ contract StorageTest {
     mapping(address => bool) public map_bool;
 
     constructor() public {
-        basic = UnpackedStruct({
-            a: 1337,
-            b: 1337
-        });
+        basic = UnpackedStruct({ a: 1337, b: 1337 });
 
-        uint256 two = (1<<128) | 1;
+        uint256 two = (1 << 128) | 1;
         map_packed[msg.sender] = two;
-        map_packed[address(bytes20(uint160(1337)))] = 1<<128;
+        map_packed[address(bytes20(uint160(1337)))] = 1 << 128;
     }
 
     function read_struct_upper(address who) public view returns (uint256) {
